+++ demo/famous.js
@@ -85,16 +85,9 @@ if (Meteor.isClient) {
     return out;
   };
 
-  /*
-  Template._yieldMain.rendered = function() {
-    var famousData = famousCmp.dataFromTpl(this);
-    famousData.viewNode.inTransformFrom(function(progress) {
-        return Transform.translate(window.innerWidth * (1.0 - progress), 0, 0);
-    });
-    famousData.viewNode.outTransformFrom(function(progress) {
-        return Transform.translate(window.innerWidth * progress - window.innerWidth, 0, 0);
-    });
+  Template.yieldMain.getTransition = function() {
+    var useForPages = Session.get('transitionPages');
+    return useForPages ? Session.get('currentTransition') : 'opacity';
   }
-  */
 
 }